requirements.txt
discord.py
loguru
PyNaCl
yt-dlp
pillow
aiohttp
python-dotenv